fixing mask, fixing install

+++ tileviewer/db/dbbuilder.py
@@ -10,33 +10,66 @@ import os
 import re
 import sys
 
-import pymongo
+try:
+    import pymongo
+except ImportError as e:
+    no_pymongo = True
 
 #cname = '140307_rep_grid'
 #dregex = 'c0_(?P<x>[-,0-9]*)_(?P<y>[-,0-9]*)_'
 dregex = '(?P<x>[-,0-9]*)_(?P<y>[-,0-9]*).tif'
 #hfov = 4150
 #ddir = '/home/graham/Desktop/140708_temcagt_bg_correction/corrected/ds/*.tif'
-ddir = '/home/graham/Desktop/montages/150707153407/processed/ds8/tiles/*.tif'
+cname = '150715144606'
+if len(sys.argv) > 1 and 'json' not in sys.argv[1]:
+    cname = sys.argv[1]
+    del sys.argv[1]
+#cname = '150715084319'
+#ddir = '/home/graham/Desktop/montages/150707153407/processed/ds8/tiles/*.tif'
+ddir = '/home/graham/montages/{}/processed/ds8/tiles/*.tif'.format(cname)
 #ddir = 'tests/tiles/*.tif'
 #cname = '140310'
-cname = '150707153407'
+#cname = '150707153407'
 #dregex = '(?P<x>[0-9]*)_(?P<y>[0-9]*)'
 #hfov = 4150
 hfov = 486
+vfov = 468
 #hfov = 0.5
+
+#ydx = 33.5
+#ydy = -56
+#xdx = 56
+#xdy = -33.6
+
+ydx = 67
+ydy = -112
+xdx = 112
+xdy = 67
+
+mask = 'image_mask.tif'
+
+
 
 
 def build_tile_spec(fn, regex):
     s = re.search(regex, fn).groupdict()
-    x = float(s['x']) * 486
-    y = -float(s['y']) * 486
+    x = (-float(s['x']) * hfov) + float(s['x']) * xdx + float(s['y']) * ydx
+    y = (float(s['y']) * vfov) + float(s['x']) * xdy + float(s['y']) * ydy
+    print fn, s, x, y
+    #x = float(s['x']) * 486
+    #y = -float(s['y']) * 486
+    if mask is not None:
+        s['mask'] = mask
     s['url'] = {"0": os.path.abspath(fn)}
     s['bbox'] = {
-        'left': x - hfov,
+        'left': x,
         'right': x + hfov,
-        'north': y + hfov,
-        'south': y - hfov,
+        'north': y + vfov,
+        'south': y,
+        #'left': x - hfov,
+        #'right': x + hfov,
+        #'north': y + hfov,
+        #'south': y - hfov,
         'top': 1000,
         'bottom': 0,
     }
@@ -73,7 +106,7 @@ def build_json(output_fn, glob_string, regex):
 
 
 if __name__ == '__main__':
-    if len(sys.argv) < 2:
+    if len(sys.argv) < 2 and not no_pymongo:
         coll = pymongo.Connection('localhost')[cname]['tiles']
         coll.drop()
         build_database(coll, ddir, dregex)
@@ -83,4 +116,8 @@ if __name__ == '__main__':
             'bbox.top', 'bbox.bottom'))
         print("Resulting database has {} tiles".format(coll.count()))
     else:
-        build_json(sys.argv[1], ddir, dregex)
+        if len(sys.argv) > 1:
+            ofn = sys.argv[1]
+        else:
+            ofn = '{}.json'.format(cname)
+        build_json(ofn, ddir, dregex)
